third_party/blink/web_tests/external/wpt/css/css-values/media-progress-computed.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] media-progress() identity check
  assert_equals: media-progress(height, 0px, 1px) and 600 serialize to the same thing in used values. expected "600" but got "1"
[FAIL] media-progress(height, media-progress(height, 0px, 1px) * 1px, media-progress(height, 0px, 1px) * 1px) should be used-value-equivalent to 0
  assert_equals: media-progress(height, media-progress(height, 0px, 1px) * 1px, media-progress(height, 0px, 1px) * 1px) and 0 serialize to the same thing in used values. expected "0" but got "1"
[FAIL] media-progress() with length product
  assert_equals: calc(media-progress(width, 0px, 50px) * 10px + 100px) and 260px serialize to the same thing in used values. expected "260px" but got "110px"
[FAIL] media-progress with complex, calculation
  assert_equals: calc(media-progress(height, 10px, sign(50px - 500em) * 10px)) and -29.5 serialize to the same thing in used values. expected "-29.5" but got "0"
[FAIL] media-progress() as length
  assert_equals: calc(media-progress(width, 0px, 1px) * 1px) and 800px serialize to the same thing in used values. expected "800px" but got "1px"
[FAIL] media-progress() as time
  assert_equals: calc(media-progress(height, 0px, 1px) * 1s) and 600s serialize to the same thing in used values. expected "600s" but got "1s"
[FAIL] media-progress() as angle
  assert_approx_equals: calc(media-progress(width, 0px, 1px) * 1deg) and 800deg 0.001 serialize to the same thing in used values. expected 800 +/- 0.001 but got 1
Harness: the test ran to completion.
